feat: Add authentication module and related files
This commit adds the authentication module and its related files. It includes the creation of the `CreateAuthDto` and `UpdateAuthDto` classes in the `auth/dto` directory. Additionally, the `User` entity is created in the `auth/entities` directory. The `AuthModule` is also added to the `imports` array in the `AppModule`.

These changes are necessary to implement authentication functionality in the application.